This commit was generated by cvs2svn to compensate for changes in r18742, which included commits to RCS files with non-trunk default branches.
git-svn-id: svn+ssh://svn.mit.edu/athena/trunk@18743 728af825-273c-0410-89f9-f7d3b574a069

+++ third/gstreamer/libs/ext/cothreads/cothreads/cothreads.c
@@ -373,9 +373,9 @@ cothread_destroy (cothread *thread)
   cothread_chunk *chunk = private->chunk;
   
 #if PTH_STACK_GROWTH > 0
-  i = (gint) ((gulong)thread - chunk->chunk) / (chunk->size / chunk->nblocks);
+  i = (gint) ((gulong)thread - (gulong)chunk->chunk) / (chunk->size / chunk->nblocks);
 #else
-  i = (gint) (chunk->chunk + chunk->size - (gulong)thread) / (chunk->size / chunk->nblocks);
+  i = (gint) ((gulong)chunk->chunk + chunk->size - (gulong)thread) / (chunk->size / chunk->nblocks);
 #endif
 
   COTHREADS_DEBUG ("destroying cothread %d (%p, %p)", i, thread, chunk->chunk);
